Move workflow node screenreader into separate file
